Updated "while" to "do-while" in getInput
I originally had while loops in getInput that required me to reset the condition after each check. I did this because my personal preference is to see the condition upfront instead of trying to find it, but the reset conditions were excess. I switched to do-while, and added the condition to the comment on the opening braces